Documentation

+++ deleteParam.js
@@ -1,13 +1,15 @@
 const execShellCommand = require("./common/execShellCommand");
 
 //---Change here or run with env variables
-const target = process.env.name || '/ns8-protect-api/pgtstblue/postgresql/password';
+const name = process.env.name || '/ns8-protect-api/pgtstblue/postgresql/password';
 //-------
 
 async function main() {
+    if(!name ) throw new Error('Please set "name" environmental variable')
+
     await execShellCommand(
         `aws ssm delete-parameter \\
-            --name ${target}`
+            --name ${name}`
     )
 }
